Lots of stuff, read carefully
Renamed fruit_critter.py to Critter.py, commented out fruit class stuff
b/c Jason uploaded a file for that seperately.
The Critter class is currently not passing a test, please feel free to
examine and run the test and then let me know if its me or the class.